requirements.txt
pytest==8.3.2
uvicorn==0.30.6
pydantic==2.8.2
fastapi==0.112.0
SQLAlchemy==2.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.9